dom/bindings/test/test_exceptions_from_jsimplemented.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=923010
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 923010</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">
  /** Test for Bug 923010 **/
  try {
    var conn = new RTCPeerConnection();
    try {
      conn.updateIce(candidate, function() {
          ok(false, "The call to updateIce succeeded when it should have thrown");
        }, function() {
          ok(false, "The call to updateIce failed when it should have thrown");
        })
      ok(false, "That call to updateIce should have thrown");
    } catch (e) {
      is(e.lineNumber, 16, "Exception should have been on line 16");
      is(e.message,
         "updateIce not yet implemented",
         "Should have the exception we expect");
    }

    var candidate = new RTCIceCandidate({candidate: null });

    conn.addIceCandidate(candidate)
    .then(function() {
      ok(false, "addIceCandidate succeeded when it should have failed");
    }, function(reason) {
      is(reason.lineNumber, 31, "Rejection should have been on line 31");
      is(reason.message,
         "Invalid candidate passed to addIceCandidate!",
         "Should have the rejection we expect");
    })
    .catch(function(reason) {
      ok(false, "unexpected error: " + reason);
    });
  } catch (e) {
    // b2g has no WebRTC, apparently
    todo(false, "No WebRTC on b2g yet");
  }
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=923010">Mozilla Bug 923010</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
